+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 9/15/2019
 
-#### Unreleased
+#### [1.1.1](https://github.com/1846689910/cicero/commit/a7e4bb34a22cf3de342a282b7984104d843884b1)
 
 ### Creates